Used x86 working build info and other open PR changes to create a minimal change macos build for testing.

+++ src/controller/src/rocprofvis_controller_trace.cpp
@@ -2555,13 +2555,13 @@ Trace::AsyncFetch(rocprofvis_property_t property, Future& future, Array& array,
                 }
                 case kRPVControllerBucketDataValueIndexed:
                 {
-                    size_t buckets_num = 0;
+                    uint64_t buckets_num = 0;
                     result = GetUInt64(kRPVControllerGetHistogramBucketsNumber, 0, &buckets_num);
                     if (result == kRocProfVisResultSuccess)
                     {
                         result = array.SetUInt64(kRPVControllerArrayNumEntries, 0, buckets_num);
                     }
-                    result = m_tracks[index]->GetBucketValues(buckets_num, array);
+                    result = m_tracks[index]->GetBucketValues(static_cast<size_t>(buckets_num), array);
                     break;
                 }
                 default:
